% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.R
\name{print.femlm.obs2remove}
\alias{print.femlm.obs2remove}
\title{Print method for femlm.obs2remove objects}
\usage{
\method{print}{femlm.obs2remove}(x, ...)
}
\arguments{
\item{x}{A \code{femlm.obs2remove} object obtained from function \code{\link[FENmlm]{obs2remove}}.}

\item{...}{Not currently used.}
}
\description{
This function show synthetizes the information of function \code{\link[FENmlm]{obs2remove}}. It reports the number of observations to be removed as well as the number of clusters removed per cluster dimension.
}
\examples{
base = iris
# v6: Petal.Length with only 0 values for 'setosa'
base$v6 = base$Petal.Length
base$v6[base$Species == "setosa"] = 0

(x = obs2remove(v6 ~ Species, base))
attr(x, "cluster")

}
